Extract EventService
