+++ app/src/main/java/net/squanchy/schedule/ScheduleModule.java
@@ -5,7 +5,7 @@ import android.content.Context;
 import net.squanchy.navigation.Navigator;
 import net.squanchy.service.firebase.FirebaseDbService;
 import net.squanchy.service.firebase.injection.DbServiceType;
-import net.squanchy.support.lang.Checksum;
+import net.squanchy.service.view.EventService;
 
 import dagger.Module;
 import dagger.Provides;
@@ -20,8 +20,8 @@ class ScheduleModule {
     }
 
     @Provides
-    ScheduleService scheduleService(@DbServiceType(DbServiceType.Type.AUTHENTICATED) FirebaseDbService dbService, Checksum checksum) {
-        return new ScheduleService(dbService, checksum);
+    ScheduleService scheduleService(@DbServiceType(DbServiceType.Type.AUTHENTICATED) FirebaseDbService dbService, EventService eventService) {
+        return new ScheduleService(dbService, eventService);
     }
 
     @Provides
